Fixed the placement of curly braces in functions and namespaces

+++ lungu.george/P2/Tylor.cpp
@@ -4,7 +4,7 @@
 
 namespace lungu
 {
-  double func_Tylor(double x, double absError, int numberMax)
+  double func_Tylor(double x, double absError, size_t numberMax)
 {
     double sum = 1.0;
     double term = 1.0;
